Add files via upload
Adapter and ViewHolder. I really tired my best to make use of epoxy but it was giving me errors and I spend way to much time on it. Hence to save time and mitigate the delay in the project I made it work the same way with the default adapter.
I still need to add pictures and descriptions. However I will do that when I get back to Dubai in three days. I will also try it with epoxy once I get back to Dubai to make it work.